assignment-3/requirements.txt
# Pytorch CPU ONLY:
accelerate==1.0.1
attrs==25.3.0
branca==0.8.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
descartes==1.1.0
diffusers==0.35.1
filelock==3.16.1
fiona==1.10.1
folium==0.18.0
fonttools==4.57.0
fsspec==2025.3.0
geopandas==0.13.2
gTTS==2.5.4
huggingface-hub==0.35.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==25.0
pandas==2.0.3
pillow==10.4.0
psutil==7.1.0
pygame==2.6.1
pyparsing==3.1.4
pyproj==3.5.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
shapely==2.0.7
six==1.17.0
sympy==1.13.3
tokenizers==0.20.3
torch==2.4.1
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
xyzservices==2025.4.0
zipp==3.20.2

# Pytorch Cuda GPU (faster image generation and better quailty)
accelerate==1.0.1
attrs==25.3.0
branca==0.8.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
descartes==1.1.0
diffusers==0.35.1
filelock==3.16.1
fiona==1.10.1
folium==0.18.0
fonttools==4.57.0
fsspec==2025.3.0
geopandas==0.13.2
gTTS==2.5.4
huggingface-hub==0.35.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==25.0
pandas==2.0.3
pillow==10.4.0
psutil==7.1.0
pygame==2.6.1
pyparsing==3.1.4
pyproj==3.5.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
shapely==2.0.7
six==1.17.0
sympy==1.13.3
tokenizers==0.20.3
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
xyzservices==2025.4.0
zipp==3.20.2
